Split out the first couple of tests from MockSupport
